fix: comparison int with size_t

+++ rust-bindings/bls-dash-sys/c-bindings/utils.hpp
@@ -24,7 +24,7 @@ template <class T>
 std::vector<T> toBLSVector(void** elems, const size_t len) {
     std::vector<T> vec;
     vec.reserve(len);
-    for (int i = 0 ; i < len; ++i) {
+    for (int i = 0 ; i < (int)len; ++i) {
         const T* el = (T*)elems[i];
         vec.push_back(*el);
     }
